\name{MLEP-package}
\alias{MLEP-package}
\alias{MLEP}
\docType{package}
\title{
Evaluate maximum likelihood estimate of penetrance parameters
}
\description{
This package provides maximum likelihood estimate of penetrance parameters for linkage analysis.
}
\details{
\tabular{ll}{
Package: \tab MLEP\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2011-12-24\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yuki Sugaya\cr
Maintainer: Yuki Sugaya <sugaya@stat.math.keio.ac.jp>
}
\references{
Sugaya Y. MLEP: Programs for exploring the maximum likelihood estimates of penetrance parameters, In preparation. 
}
\keyword{ package }


